修改readme

+++ sh/os_cmd.py
@@ -1,7 +1,7 @@
 
 import shutil
 import os
-
+from sh.utils import normal_leven
 WORK_DIR = ""
 """
  param: f: files
@@ -9,7 +9,7 @@ WORK_DIR = ""
 """
 
 def is_file(filename):
-    return not os.path.isdir(filename)
+    return os.path.isfile(filename)
 
 def is_dir(path):
     return os.path.isdir(path)
@@ -54,16 +54,40 @@ def mkdir(path):
     os.mkdir(path)
 
 def rm(fpath, p=""):
+    if len(fpath) < 2: return
     if "r" not in p:
         os.remove(fpath)
     else:
         shutil.rmtree(fpath)
 
-def cp():
-    pass
+def cptree(src, dst, symlinks=False, ignore=None):
+    for item in os.listdir(src):
+        s = os.path.join(src, item)
+        d = os.path.join(dst, item)
+        if os.path.isdir(s):
+            shutil.copytree(s, d, symlinks, ignore)
+        else:
+            shutil.copy2(s, d)
 
-def mv():
-    pass
+def cp(src, dst, p=""):
+    if "r" in p and is_dir(src):
+        cptree(src, dst)
+    else:
+        shutil.copy2(src, dst)
+
+def mv(src, dst):
+    shutil.move(src, dst)
+
 
 def find(name, dir_path="." , p="r"):
-    pass
+    files = ls(dir_path, p)
+    ans = []
+    for fn in files:
+        head, tail = os.path.split(dir_path)
+        dis = normal_leven(tail, name)
+        ans.append((dis, fn))
+    sorted(ans)
+    ans = ans[0:5]
+    if len(ans) > 0 and ans[0][1] == 0:
+        return [ans[0][1]]
+    return list(map(lambda x:x[1], ans))
